Move streamer test to integration test
Failing in the coverage build.
